fix: ga 연결 방식 변경
