+++ lib/gtag.js
@@ -1,17 +1,15 @@
 export const GA_TRACKING_ID = process.env.NEXT_PUBLIC_GA_ID;
 
-// https://developers.google.com/analytics/devguides/collection/gtagjs/pages
 export const pageview = (url) => {
   window.gtag('config', GA_TRACKING_ID, {
     page_path: url,
   });
 };
 
-// https://developers.google.com/analytics/devguides/collection/gtagjs/events
 export const event = ({ action, category, label, value }) => {
   window.gtag('event', action, {
     event_category: category,
     event_label: label,
-    value: value,
+    value,
   });
 };
