+++ pages/_app.tsx
@@ -5,15 +5,16 @@ import { useEffect } from 'react';
 
 import Layout from '../components/Layout';
 import '../styles/globals.css';
-import * as ga from '../lib/ga';
+import * as gtag from '../lib/gtag';
 
 export default function App({ Component, pageProps }: AppProps) {
   const router = useRouter();
 
   useEffect(() => {
-    const handleRouteChange = (url: unknown) => {
-      ga.pageview(url);
+    const handleRouteChange = (url) => {
+      gtag.pageview(url);
     };
+
     router.events.on('routeChangeComplete', handleRouteChange);
 
     return () => {
@@ -22,8 +23,28 @@ export default function App({ Component, pageProps }: AppProps) {
   }, [router.events]);
 
   return (
-    <Layout>
-      <Component {...pageProps} />
-    </Layout>
+    <>
+      <Script
+        strategy="afterInteractive"
+        src={`https://www.googletagmanager.com/gtag/js?id=${process.env.NEXT_PUBLIC_GA_ID}`}
+      />
+      <Script
+        id="google-analytics"
+        strategy="afterInteractive"
+        dangerouslySetInnerHTML={{
+          __html: `
+          window.dataLayer = window.dataLayer || [];
+          function gtag(){dataLayer.push(arguments);}
+          gtag('js', new Date());
+          gtag('config', ${process.env.NEXT_PUBLIC_GA_ID}, {
+            page_path: window.location.pathname,
+          });
+        `,
+        }}
+      />
+      <Layout>
+        <Component {...pageProps} />
+      </Layout>
+    </>
   );
 }
